DOCO: Added information about usage of HTMLField
Spent quite a bit of time trying to figure out why additional plugins were showing up inside the text editor when using HTMLField. Found out it isn't supported when using this field in this issue:

https://github.com/divio/djangocms-text-ckeditor/issues/72

Decided that I'm not the first person to run into this question, so I'm documenting it for future reference.