Fixed block level plugins wrapped with paragraphs breaking markup
Previous versions of the addon assumed that every plugin is an inline
plugin, which could result in the block level markup wrapped with a `<p>` tag
With new system in place it is no longer allowed, but previous markup
could potentially be broken on first edit. This commit prevents it.

Ref: #353