Fixed a bug with disappearing whitespace before "inline" plugins
We were setting the `cms-plugin` tag as a nonEditable but that's not
really needed since the widgets are non-editable anyway. What it did was
setting "isBlockLike" property on ckeditor html parser elements, which
in turn trimmed whitespace on the right of any text node preceding this
`cms-plugin`, which would result in e.g. links sticking to the text
before them.